Refactor code

+++ utilities/utilities.go
@@ -7,14 +7,8 @@ import (
 	"time"
 )
 
-//def print_progress_bar(index, total_rows)
-//progress = index.to_f / total_rows
-//print '[' + ('*' * (progress * 50)) + (' ' * ((1 - progress) * 50)) + ']'
-//print "\r"
-//end
-
-func PrintProgressBar(index, totalRows float64, totalTime time.Duration, tail string) {
-	progress := index / totalRows
+func PrintProgressBar(index, totalRows int, totalTime time.Duration, tail string) {
+	progress := float64(index) / float64(totalRows)
 	bar := strings.Repeat("=", int(math.Round(progress*50.0)))
 	space := strings.Repeat(" ", int(math.Round((1-progress)*50)))
 	averageTime := AverageTime(totalTime, index)
@@ -22,8 +16,8 @@ func PrintProgressBar(index, totalRows float64, totalTime time.Duration, tail st
 	fmt.Print("\r")
 }
 
-func AverageTime(totalTime time.Duration, index float64) time.Duration {
-	return time.Duration(float64(totalTime) / index)
+func AverageTime(totalTime time.Duration, index int) time.Duration {
+	return time.Duration(float64(totalTime) / float64(index))
 }
 
 func Percent(num int, denom int) int {
